Support 'pass' service and extended service names
Services in master.cf can be of different types, one of them being
'pass', which was not supported. Furthermore, different types have
different name formats. The 'inet' type allows host:port to be
specified in various formats, while all other types allow a
path/to/a/socket as the name.
This change relaxes the validation and matching patterns to allow such
extended names.
NOTE: With inet service names being able to contain semicolons, the
config_dir must now always be specified explicitly (because any prefix
will be matched to the service property)