Make fact lenient if postfix is not installed
If postfix is not installed, the postfix fact should simply not have any
value at all. Before this change, it would cause an error message
because the postconf binary was not found.
The confine statement in place (and moved by the previous fix for this
issue from pull request [#6]) did not solve the problem, because it used
Puppet confine syntax/features not compatible with Facter.

[#6]: https://github.com/oxc/puppet-postfix/pull/6