Remove postfix_master confine on EOL postfix version
The binary used by the postfix_master provider was introduced only in
postfix v2.11, which is why the provider was confined to this version.

However, this causes trouble on initial run, when the package is not yet
installed. Since postfix 2.11 is already EOL anyway, the easiest fix is
to removed the confine.

Patch provided by Craig Watson (see #10).

Fixes #9.