update documentation regarding the `config_dir` parameter
In 3065f1615d8cceeb78b07282bebb122dd572e86f this parameter was removed
from all types. Instead it is now extracted from the resource title.
If the resource title starts with `foo::`, then 'foo' will be added
to Postfix's configuration path. For example, if the configuration
path is /etc/postfix, it will then be changed to /etc/postfix-foo for
this resource.