fix assorted multi instance issues
First, fix support for specifying custom instance directories.
Previously a config parameter, that was supposed to be added to a
*custom* instance, could be added to the *main* instance of Postfix
instead. This is wrong and should never happen.

This occured when the provider was unable to get the config
directory from `postmulti -l`. Now it is ensured that the custom
directory is used by composing the path in the provider if necessary.

The tests actually revealed this, but they were mostly nonfunctional
and could not be run. So this was only discovered just yet.

Additionally creating new instances using postmulti was broken.
A new instance could not be created if multi instance support was not
already initialized in Postfix. The provider now ensures that this is
the case by always running the 'init' command prior to creating a new
instance.

A few other changes in this commit:
* fix acceptance tests for containers with misconfigured networking
* replace outdated syntax in acceptance tests